titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/cn-validation/data-access-cn-validation/src/lib/+state/cn-validation.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectCnValidationState } from './cn-validation.reducer';

const selectCnValidationList = createSelector(
  selectCnValidationState,
  state => state.cnValidationList
);

const selectCnTypeList = createSelector(
  selectCnValidationState,
  state => state.cnTypeList
);

const selectCnValidation = createSelector(
  selectCnValidationState,
  state => state.cnValidation
);

const selectError = createSelector(
  selectCnValidationState,
  state => state.error
);
const selectIsLoading = createSelector(
  selectCnValidationState,
  state => state.isLoading
);
const selectHassaved = createSelector(
  selectCnValidationState,
  state => state.hasSaved
);
const selectHasUpdated = createSelector(
  selectCnValidationState,
  state => state.hasUpdated
);
const selectTotalElement = createSelector(
  selectCnValidationState,
  state => state.totalElements
);

export const cnValidationSelectors = {
  selectCnValidationList,
  selectCnValidation,
  selectError,
  selectHasUpdated,
  selectHassaved,
  selectIsLoading,
  selectTotalElement,
  selectCnTypeList
};
